2.17.5

Conflicts:
	CMakeLists.txt

+++ src/storaged/sconfig.c
@@ -366,7 +366,7 @@ int sconfig_read(sconfig_t *config, const char *fname, int cluster_id) {
         }
         else {
           if (strlen(spare_mark) > 9) {
-            severe("cid%d/sid%d has too long spare-mark : strlen(%s) = %d >9.", cid, sid, spare_mark, (int)strlen(spare_mark));
+            severe("cid%d/sid%d has too long spare-mark : strlen(%s) = %d >9.", (int)cid, (int)sid, spare_mark, (int)strlen(spare_mark));
           }
         }
